Renamed folders and refactored code
