+++ gpt/model.py
@@ -1,10 +1,8 @@
-import math
-
 import torch
-from torch import nn
 import torch.nn.functional as F
+from torch import nn
 
-from modeling.config import GPTConfig
+from gpt.config import GPTConfig
 
 
 class GPT(nn.Module):
