+++ requirements.txt
@@ -1,0 +1,26 @@
+certifi==2022.12.7
+charset-normalizer==2.1.1
+colorama==0.4.6
+filelock==3.9.0
+idna==3.4
+iniconfig==2.0.0
+Jinja2==3.1.2
+MarkupSafe==2.1.2
+mpmath==1.3.0
+networkx==3.0
+numpy==1.24.1
+packaging==23.1
+Pillow==9.3.0
+pip==23.2.1
+pluggy==1.3.0
+pytest==7.4.2
+requests==2.31.0
+setuptools==68.2.2
+sympy==1.12
+--find-links https://download.pytorch.org/whl/torch_stable.html
+torch==2.0.1+cu118
+torchaudio==2.0.2+cu118
+torchvision==0.15.2+cu118
+typing-extensions==4.8.0
+urllib3==2.0.5
+wheel==0.41.2
