+++ tests/test_gpt/test_config.py
@@ -1,5 +1,5 @@
 # Generated by CodiumAI
-from modeling.config import GPTConfig
+from gpt.config import GPTConfig
 
 
 class TestGPTConfig:
